+++ web/components/songcard/index.tsx
@@ -1,0 +1,2 @@
+export * from "./SongCard"
+export * from "./MoreSongCard"
