+++ web/components/u/index.tsx
@@ -1,0 +1,2 @@
+export * from "./types"
+export * from "./U"
