+++ web/components/u/types.ts
@@ -1,0 +1,7 @@
+export enum DetailKind {
+  Song, Album
+}
+
+export enum Mode {
+  All, Loved, Color
+}
